Fix the main widget tests
These tests now properly show the app and make sure that the events are
triggered in the same manner as when a person is using the widget. The
main part here is that now the return key test fails which is what we
were trying to test in the first place.